# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Medical Appointment System API
  description: |
    API para el sistema de agendamiento de citas médicas para asegurados.
    
    ## Flujo del Sistema
    1. El cliente crea una solicitud de cita (POST /appointments)
    2. La solicitud se guarda en DynamoDB con estado "pending"
    3. Se publica un mensaje a SNS según el país (PE o CL)
    4. Una lambda específica por país procesa la cita
    5. Se guarda en la base de datos RDS del país correspondiente
    6. Se publica un evento de confirmación a EventBridge
    7. Se actualiza el estado en DynamoDB a "completed"
    8. El cliente puede consultar el estado (GET /appointments/{insuredId})
    
  version: 1.0.0
  contact:
    name: Backend Team
    email: backend@medical.com
  license:
    name: MIT
servers:
  - url: https://api.medical.com/dev
    description: Development server
  - url: https://api.medical.com/prod
    description: Production server

tags:
  - name: Appointments
    description: Operaciones relacionadas con citas médicas

paths:
  /appointments:
    post:
      tags:
        - Appointments
      summary: Crear nueva solicitud de cita médica
      description: |
        Crea una nueva solicitud de cita médica. La solicitud se procesa de forma asíncrona.
        El sistema devuelve inmediatamente con el ID de la cita y estado "pending".
      operationId: createAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentRequest'
            examples:
              peru:
                summary: Cita para Perú
                value:
                  insuredId: "00123"
                  scheduleId: 100
                  countryISO: "PE"
              chile:
                summary: Cita para Chile
                value:
                  insuredId: "99999"
                  scheduleId: 200
                  countryISO: "CL"
      responses:
        '201':
          description: Solicitud de cita creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAppointmentResponse'
              example:
                success: true
                message: "Appointment created successfully"
                data:
                  appointmentId: "550e8400-e29b-41d4-a716-446655440000"
                  status: "pending"
                  message: "Appointment request is being processed"
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "Bad Request"
                error: "insuredId must be a 5-digit string"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /appointments/{insuredId}:
    get:
      tags:
        - Appointments
      summary: Obtener citas por ID de asegurado
      description: |
        Retorna todas las citas médicas asociadas a un asegurado específico,
        incluyendo su estado actual (pending, completed, failed).
      operationId: getAppointmentsByInsuredId
      parameters:
        - name: insuredId
          in: path
          required: true
          description: ID del asegurado (5 dígitos)
          schema:
            type: string
            pattern: '^\d{5}$'
            example: "00123"
      responses:
        '200':
          description: Lista de citas obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAppointmentsResponse'
              example:
                success: true
                message: "Appointments retrieved successfully"
                data:
                  insuredId: "00123"
                  count: 2
                  appointments:
                    - appointmentId: "550e8400-e29b-41d4-a716-446655440000"
                      scheduleId: 100
                      countryISO: "PE"
                      status: "completed"
                      createdAt: "2024-12-01T10:00:00Z"
                      updatedAt: "2024-12-01T10:05:00Z"
                    - appointmentId: "660e8400-e29b-41d4-a716-446655440001"
                      scheduleId: 101
                      countryISO: "PE"
                      status: "pending"
                      createdAt: "2024-12-02T14:30:00Z"
                      updatedAt: "2024-12-02T14:30:00Z"
        '400':
          description: ID de asegurado inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No se encontraron citas para el asegurado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CreateAppointmentRequest:
      type: object
      required:
        - insuredId
        - scheduleId
        - countryISO
      properties:
        insuredId:
          type: string
          pattern: '^\d{5}$'
          description: ID del asegurado (5 dígitos, puede incluir ceros al inicio)
          example: "00123"
        scheduleId:
          type: integer
          minimum: 1
          description: ID del espacio de cita (centro médico + especialidad + médico + fecha/hora)
          example: 100
        countryISO:
          type: string
          enum: [PE, CL]
          description: Código ISO del país (PE para Perú, CL para Chile)
          example: "PE"

    CreateAppointmentResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Appointment created successfully"
        data:
          type: object
          properties:
            appointmentId:
              type: string
              format: uuid
              description: ID único de la cita generado por el sistema
              example: "550e8400-e29b-41d4-a716-446655440000"
            status:
              type: string
              enum: [pending, completed, failed]
              example: "pending"
            message:
              type: string
              example: "Appointment request is being processed"

    GetAppointmentsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Appointments retrieved successfully"
        data:
          type: object
          properties:
            insuredId:
              type: string
              example: "00123"
            count:
              type: integer
              example: 2
            appointments:
              type: array
              items:
                $ref: '#/components/schemas/Appointment'

    Appointment:
      type: object
      properties:
        appointmentId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        scheduleId:
          type: integer
          example: 100
        countryISO:
          type: string
          enum: [PE, CL]
          example: "PE"
        status:
          type: string
          enum: [pending, completed, failed]
          description: |
            - pending: La cita está siendo procesada
            - completed: La cita fue agendada exitosamente
            - failed: Ocurrió un error al procesar la cita
          example: "completed"
        createdAt:
          type: string
          format: date-time
          example: "2024-12-01T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-12-01T10:05:00Z"
        errorMessage:
          type: string
          nullable: true
          description: Mensaje de error (solo presente si status = failed)
          example: null

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Bad Request"
        error:
          type: string
          example: "insuredId must be a 5-digit string"